Feature user management (#30)
* CRUD All Users, Rating

* UPDATE logic and fix class convention

* FIX role and permission

* REFACTOR convention, return object at save and update function

+++ src/main/java/com/crm/models/Bid.java
@@ -59,7 +59,7 @@ public class Bid {
   private Set<Container> containers = new HashSet<>();
 
   @Column(name = "bid_price")
-  private float bidPrice;
+  private Double bidPrice;
 
   @Column(name = "bid_date")
   private LocalDateTime bidDate;
